# ai-resume-match/api/openapi.json
{
  "components" : {
    "schemas" : {
...
      "AnalysisPageResponse" : {
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/AnalysisListItemResponse"
            },
            "type" : "array"
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "AnalysisSummaryResponse" : {
        "properties" : {
          "averageMatchScore" : {
            "type" : "number"
          },
          "inProgressCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "retryableFailureCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "successCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalCount" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "AnalysisTaskResponse" : {
        "properties" : {
          "attemptCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "completedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "failureCode" : {
            "type" : "string"
          },
          "failureMessage" : {
            "type" : "string"
          },
          "jobDescriptionId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "jobTitle" : {
            "type" : "string"
          },
          "matchScore" : {
            "format" : "int32",
            "type" : "integer"
          },
          "maxAttempts" : {
            "format" : "int32",
            "type" : "integer"
          },
          "nextRetryAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "resumeFileName" : {
            "type" : "string"
          },
          "resumeId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "startedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "taskId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ApiProblemDetail" : {
        "description" : "RFC 9457 problem details with stable application error code and request ID",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "detail" : {
            "type" : "string"
          },
          "instance" : {
            "format" : "uri",
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "requestId" : {
            "nullable" : true,
            "type" : "string"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "title" : {
            "type" : "string"
          },
          "type" : {
            "format" : "uri",
            "type" : "string"
          }
        },
        "required" : [ "code", "detail", "instance", "message", "requestId", "status", "title", "type" ],
        "type" : "object"
      },
      "CreateAnalysisRequest" : {
        "properties" : {
          "jobDescriptionId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "resumeId" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "jobDescriptionId", "resumeId" ],
        "type" : "object"
      },
      "CreateJobRequest" : {
        "properties" : {
          "content" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "JobDescriptionResponse" : {
        "properties" : {
          "jobDescriptionId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "title" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "MatchReportResponse" : {
        "properties" : {
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "matchScore" : {
            "format" : "int32",
            "maximum" : 100,
            "minimum" : 0,
            "type" : "integer"
          },
          "provenance" : {
            "$ref" : "#/components/schemas/AnalysisProvenance"
          },
          "reportContent" : {
            "type" : "string"
          },
          "reportSchemaVersion" : {
            "enum" : [ "markdown-v1", "match-report-v2" ],
            "type" : "string"
          },
          "structuredReport" : {
            "$ref" : "#/components/schemas/StructuredMatchReport"
          },
          "taskId" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "createdAt", "matchScore", "provenance", "reportContent", "reportSchemaVersion", "structuredReport", "taskId" ],
        "type" : "object"
      },
...
      "ResumeUploadResponse" : {
        "properties" : {
          "resumeId" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "apiToken" : {
        "description" : "Application API token",
        "in" : "header",
        "name" : "X-API-Token",
        "type" : "apiKey"
      }
    }
  },
  "info" : {
    "description" : "Asynchronous resume and job matching API",
    "title" : "AI Resume Match API",
    "version" : "v1"
  },
  "openapi" : "3.0.1",
  "paths" : {
    "/api/analysis" : {
      "get" : {
        "operationId" : "analyses",
        "parameters" : [ {
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "enum" : [ "PENDING", "RUNNING", "SUCCESS", "FAILED_RETRYABLE", "FAILED_FINAL", "CANCELLED", "FAILED" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AnalysisPageResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemDetail"
                }
              }
            },
            "description" : "Invalid request"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemDetail"
                }
              }
            },
            "description" : "Missing or invalid API token"
          }
        },
        "security" : [ {
          "apiToken" : [ ]
        } ],
        "tags" : [ "resume-match-controller" ]
      },
      "post" : {
        "operationId" : "createAnalysis",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateAnalysisRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "202" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AnalysisTaskResponse"
                }
              }
            },
            "description" : "Analysis accepted",
            "headers" : {
              "Location" : {
                "description" : "Canonical task status resource",
                "schema" : {
                  "type" : "string"
                },
                "style" : "simple"
              }
            }
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemDetail"
                }
              }
            },
            "description" : "Invalid request"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemDetail"
                }
              }
            },
            "description" : "Missing or invalid API token"
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemDetail"
                }
              }
            },
            "description" : "Anonymous session analysis quota exceeded; see Retry-After"
          },
          "503" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemDetail"
                }
              }
            },
            "description" : "Analysis quota enforcement is temporarily unavailable"
          }
        },
        "security" : [ {
          "apiToken" : [ ]
        } ],
        "summary" : "Create an asynchronous analysis task from existing resources",
        "tags" : [ "resume-match-controller" ]
      }
    },
    "/api/analysis-submissions" : {
      "post" : {
        "operationId" : "createAnalysisSubmission",
        "parameters" : [ {
          "in" : "query",
          "name" : "jobTitle",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "jobContent",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : false,
          "schema" : {
            "maxLength" : 128,
            "minLength" : 0,
            "pattern" : "[A-Za-z0-9][A-Za-z0-9._:-]{0,127}",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "properties" : {
                  "file" : {
                    "format" : "binary",
                    "type" : "string"
                  }
                },
                "required" : [ "file" ],
                "type" : "object"
              }
            }
          }
        },
        "responses" : {
          "202" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AnalysisTaskResponse"
                }
              }
            },
            "description" : "Analysis accepted",
            "headers" : {
              "Location" : {
                "description" : "Canonical task status resource",
                "schema" : {
                  "type" : "string"
                },
                "style" : "simple"
              }
            }
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemDetail"
                }
              }
            },
            "description" : "Invalid request"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemDetail"
                }
              }
            },
            "description" : "Missing or invalid API token"
          },
          "409" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemDetail"
                }
              }
            },
            "description" : "Idempotency key conflict"
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemDetail"
                }
              }
            },
            "description" : "Anonymous session analysis quota exceeded; see Retry-After"
          },
          "503" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemDetail"
                }
              }
            },
            "description" : "Analysis quota enforcement is temporarily unavailable"
          }
        },
        "security" : [ {
          "apiToken" : [ ]
        } ],
        "summary" : "Atomically upload a resume, create a job and accept an analysis task",
        "tags" : [ "resume-match-controller" ]
      }
    },
    "/api/analysis/summary" : {
      "get" : {
        "operationId" : "analysisSummary",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AnalysisSummaryResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemDetail"
                }
              }
            },
            "description" : "Invalid request"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemDetail"
                }
              }
            },
            "description" : "Missing or invalid API token"
          }
        },
        "security" : [ {
          "apiToken" : [ ]
        } ],
        "tags" : [ "resume-match-controller" ]
      }
    },
    "/api/analysis/{taskId}" : {
      "delete" : {
        "operationId" : "deleteAnalysis",
        "parameters" : [ {
          "in" : "path",
          "name" : "taskId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Analysis data deleted"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemDetail"
                }
              }
            },
            "description" : "Invalid request"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemDetail"
                }
              }
            },
            "description" : "Missing or invalid API token"
          },
          "404" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemDetail"
                }
              }
            },
            "description" : "Analysis does not exist or belongs to another session"
          }
        },
        "security" : [ {
          "apiToken" : [ ]
        } ],
        "summary" : "Delete an owned analysis and its source data",
        "tags" : [ "resume-match-controller" ]
      },
      "get" : {
        "operationId" : "analysisTask",
        "parameters" : [ {
          "in" : "path",
          "name" : "taskId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AnalysisTaskResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemDetail"
                }
              }
            },
            "description" : "Invalid request"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemDetail"
                }
              }
            },
            "description" : "Missing or invalid API token"
          }
        },
        "security" : [ {
          "apiToken" : [ ]
        } ],
        "tags" : [ "resume-match-controller" ]
      }
    },
    "/api/analysis/{taskId}/report" : {
      "get" : {
        "operationId" : "report",
        "parameters" : [ {
          "in" : "path",
          "name" : "taskId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MatchReportResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemDetail"
                }
              }
            },
            "description" : "Invalid request"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemDetail"
                }
              }
            },
            "description" : "Missing or invalid API token"
          }
        },
        "security" : [ {
          "apiToken" : [ ]
        } ],
        "tags" : [ "resume-match-controller" ]
      }
    },
    "/api/analysis/{taskId}/retry" : {
      "post" : {
        "operationId" : "retryAnalysis",
        "parameters" : [ {
          "in" : "path",
          "name" : "taskId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AnalysisTaskResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemDetail"
                }
              }
            },
            "description" : "Invalid request"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemDetail"
                }
              }
            },
            "description" : "Missing or invalid API token"
          },
          "429" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemDetail"
                }
              }
            },
            "description" : "Anonymous session analysis quota exceeded; see Retry-After"
          },
          "503" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemDetail"
                }
              }
            },
            "description" : "Analysis quota enforcement is temporarily unavailable"
          }
        },
        "security" : [ {
          "apiToken" : [ ]
        } ],
        "tags" : [ "resume-match-controller" ]
      }
    },
    "/api/jobs" : {
      "post" : {
        "operationId" : "createJob",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateJobRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/JobDescriptionResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemDetail"
                }
              }
            },
            "description" : "Invalid request"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemDetail"
                }
              }
            },
            "description" : "Missing or invalid API token"
          }
        },
        "security" : [ {
          "apiToken" : [ ]
        } ],
        "tags" : [ "resume-match-controller" ]
      }
    },
    "/api/resumes" : {
      "post" : {
        "operationId" : "uploadResume",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "properties" : {
                  "file" : {
                    "format" : "binary",
                    "type" : "string"
                  }
                },
                "required" : [ "file" ],
                "type" : "object"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResumeUploadResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemDetail"
                }
              }
            },
            "description" : "Invalid request"
          },
          "401" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiProblemDetail"
                }
              }
            },
            "description" : "Missing or invalid API token"
          }
        },
        "security" : [ {
          "apiToken" : [ ]
        } ],
        "tags" : [ "resume-match-controller" ]
      }
    }
  }
}